+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.32.15](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.14...@coveo/quantic@0.32.15) (2021-07-09)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.14](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.13...@coveo/quantic@0.32.14) (2021-07-08)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.13](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.12...@coveo/quantic@0.32.13) (2021-07-08)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [0.32.12](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.11...@coveo/quantic@0.32.12) (2021-07-08)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [0.32.11](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.32.10...@coveo/quantic@0.32.11) (2021-07-06)
 
 **Note:** Version bump only for package @coveo/quantic