+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.36.2](https://github.com/coveo/ui-kit/compare/headless-react@0.36.1...headless-react@0.36.2) (2021-07-09)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.36.1](https://github.com/coveo/ui-kit/compare/headless-react@0.36.0...headless-react@0.36.1) (2021-07-08)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.36.0](https://github.com/coveo/ui-kit/compare/headless-react@0.35.4...headless-react@0.36.0) (2021-07-08)
+
+
+### Features
+
+* **headless:** execute Triggers ([#944](https://github.com/coveo/ui-kit/issues/944)) ([f429f1c](https://github.com/coveo/ui-kit/commit/f429f1cc64cd362358a7273b1f551026d1f5a74a))
+
+
+
+
+
+## [0.35.4](https://github.com/coveo/ui-kit/compare/headless-react@0.35.3...headless-react@0.35.4) (2021-07-08)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 ## [0.35.3](https://github.com/coveo/ui-kit/compare/headless-react@0.35.2...headless-react@0.35.3) (2021-07-06)
 
 **Note:** Version bump only for package headless-react